+++ guillotina/configure/config.py
@@ -20,8 +20,6 @@ from guillotina.schema import ValidationError
 from keyword import iskeyword
 from zope.interface import implementer
 from zope.interface import Interface
-from zope.interface import providedBy
-from zope.interface.adapter import AdapterRegistry
 
 import operator
 import sys
@@ -41,30 +39,6 @@ _import_chickens = {}, {}, ("*",)  # dead chickens needed by __import__
 class ConfigurationContext(object):
     """Mix-in that implements IConfigurationContext
 
-    Subclasses provide a ``package`` attribute and a ``basepath``
-    attribute.  If the base path is not None, relative paths are
-    converted to absolute paths using the the base path. If the
-    package is not none, relative imports are performed relative to
-    the package.
-
-    In general, the basepath and package attributes should be
-    consistent. When a package is provided, the base path should be
-    set to the path of the package directory.
-
-    Subclasses also provide an ``actions`` attribute, which is a list
-    of actions, an ``includepath`` attribute, and an ``info``
-    attribute.
-
-    The include path is appended to each action and is used when
-    resolving conflicts among actions.  Normally, only the a
-    ConfigurationMachine provides the actions attribute. Decorators
-    simply use the actions of the context they decorate. The
-    ``includepath`` attribute is a tuple of names.  Each name is
-    typically the name of an included configuration file.
-
-    The ``info`` attribute contains descriptive information helpful
-    when reporting errors.  If not set, it defaults to an empty string.
-
     The actions attribute is a sequence of dictionaries where each dictionary
     has the following keys:
 
@@ -77,18 +51,21 @@ class ConfigurationContext(object):
 
       - ``kw``, keyword arguments for the action
 
-      - ``includepath``, a tuple of include file names (defaults to ())
-
-      - ``info``, an object that has descriptive information about
-        the action (defaults to '')
-
     """
 
     def __init__(self):
         super(ConfigurationContext, self).__init__()
+        # on commit, we increment so we can override registered components
+        self.count = 0
+        self.module_name = 'guillotina'
 
-    def action(self, discriminator, callable=None, args=(), kw=None, order=0,
-               includepath=None, info=None, **extra):
+    def begin(self, module_name):
+        self.module_name = module_name
+
+    def commit(self):
+        self.count += 1
+
+    def action(self, discriminator, callable=None, args=(), kw=None, **extra):
         """Add an action with the given discriminator, callable and arguments.
 
         For testing purposes, the callable and arguments may be omitted.
@@ -102,95 +79,26 @@ class ConfigurationContext(object):
 
         action = extra
 
-        if info is None:
-            info = getattr(self, 'info', '')
-
-        if includepath is None:
-            includepath = getattr(self, 'includepath', ())
-
         action.update(dict(
             discriminator=discriminator,
             callable=callable,
             args=args,
             kw=kw,
-            includepath=includepath,
-            info=info,
-            order=order))
+            module_name=self.module_name,
+            order=self.count))
 
         self.actions.append(action)
 
 
-class ConfigurationAdapterRegistry(object):
-    """Simple adapter registry that manages directives as adapters
-    """
-
-    def __init__(self):
-        super(ConfigurationAdapterRegistry, self).__init__()
-        self._registry = {}
-        # Stores tuples of form:
-        #   (namespace, name), schema, usedIn, info, parent
-        self._doc_registry = []
-
-    def register(self, interface, name, factory):
-        r = self._registry.get(name)
-        if r is None:
-            r = AdapterRegistry()
-            self._registry[name] = r
-
-        r.register([interface], Interface, '', factory)
-
-    def factory(self, context, name):
-        r = self._registry.get(name)
-        if r is None:
-            # Try namespace-independent name
-            ns, n = name
-            r = self._registry.get(n)
-            if r is None:
-                raise ConfigurationError("Unknown directive", ns, n)
-
-        f = r.lookup1(providedBy(context), Interface)
-        if f is None:
-            raise ConfigurationError(
-                "The directive %s cannot be used in this context" % (name, ))
-        return f
-
-
 @implementer(IConfigurationContext)
-class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
+class ConfigurationMachine(ConfigurationContext):
     """Configuration machine
     """
-    package = None
-    basepath = None
-    includepath = ()
-    info = ''
 
     def __init__(self):
         super(ConfigurationMachine, self).__init__()
         self.actions = []
         self.stack = [RootStackItem(self)]
-        self.i18n_strings = {}
-
-    def begin(self, __name, __data=None, __info=None, **kw):
-        if __data:
-            if kw:
-                raise TypeError("Can't provide a mapping object and keyword "
-                                "arguments")
-        else:
-            __data = kw
-        self.stack.append(self.stack[-1].contained(__name, __data, __info))
-
-    def end(self):
-        self.stack.pop().finish()
-
-    def __call__(self, __name, __info=None, **__kw):
-        self.begin(__name, __kw, __info)
-        self.end()
-
-    def get_info(self):
-        return self.stack[-1].context.info
-
-    def set_info(self, info):
-        self.stack[-1].context.info = info
 
     def execute_actions(self, clear=True, testing=False):
         """Execute the configuration actions.
@@ -204,7 +112,6 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
                     continue
                 args = action['args']
                 kw = action['kw']
-                info = action['info']
                 try:
                     callable(*args, **kw)
                 except (KeyboardInterrupt, SystemExit):  # pragma NO COVER
@@ -214,7 +121,7 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
                         raise
                     t, v, tb = sys.exc_info()
                     try:
-                        reraise(ConfigurationExecutionError(t, v, info),
+                        reraise(ConfigurationExecutionError(t, v),
                                 None, tb)
                     finally:
                         del t, v, tb
@@ -227,11 +134,11 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
 class ConfigurationExecutionError(ConfigurationError):
     """An error occurred during execution of a configuration action
     """
-    def __init__(self, etype, evalue, info):
-        self.etype, self.evalue, self.info = etype, evalue, info
+    def __init__(self, etype, evalue):
+        self.etype, self.evalue = etype, evalue
 
     def __str__(self):  # pragma NO COVER
-        return "%s: %s\n  in:\n  %s" % (self.etype, self.evalue, self.info)
+        return "%s: %s\n  in:\n  %s" % (self.etype, self.evalue)
 
 
 ##############################################################################
@@ -245,14 +152,11 @@ class IStackItem(Interface):
     A stack item is created for each directive use.
     """
 
-    def contained(name, data, info):
+    def contained(name, data):
         """Begin processing a contained directive
 
         The data are a dictionary of attribute names mapped to unicode
         strings.
-
-        The info argument is an object that can be converted to a
-        string and that contains information about the directive.
 
         The begin method returns the next item to be placed on the stack.
         """
@@ -274,14 +178,13 @@ class SimpleStackItem(object):
     has been reached.
     """
     # XXX why this *argdata hack instead of schema, data?
-    def __init__(self, context, handler, info, *argdata):
+    def __init__(self, context, handler, *argdata):
         newcontext = GroupingContextDecorator(context)
-        newcontext.info = info
         self.context = newcontext
         self.handler = handler
         self.argdata = argdata
 
-    def contained(self, name, data, info):
+    def contained(self, name, data):
         raise ConfigurationError("Invalid directive %s" % str(name))
 
     def finish(self):
@@ -306,7 +209,7 @@ class RootStackItem(object):
     def __init__(self, context):
         self.context = context
 
-    def contained(self, name, data, info):
+    def contained(self, name, data):
         """Handle a contained directive
 
         We have to compute a new stack item by getting a named adapter
@@ -315,7 +218,7 @@ class RootStackItem(object):
         factory = self.context.factory(self.context, name)
         if factory is None:
             raise ConfigurationError("Invalid directive", name)
-        adapter = factory(self.context, data, info)
+        adapter = factory(self.context, data)
         return adapter
 
     def finish(self):
@@ -347,9 +250,9 @@ class GroupingStackItem(RootStackItem):
                 self.context.action(**action)
         self.__call_before = noop
 
-    def contained(self, name, data, info):
+    def contained(self, name, data):
         self.__call_before()
-        return RootStackItem.contained(self, name, data, info)
+        return RootStackItem.contained(self, name, data)
 
     def finish(self):
         self.__call_before()
@@ -363,60 +266,6 @@ class GroupingStackItem(RootStackItem):
 
 def noop():
     pass
-
-
-@implementer(IStackItem)
-class ComplexStackItem(object):
-    """Complex stack item
-
-    A complex stack item is in the stack when a complex directive is
-    being processed.  It only allows subdirectives to be used.
-
-    A complex stack item is created with a complex directive
-    definition (IComplexDirectiveContext), a configuration context,
-    and directive data.
-    """
-    def __init__(self, meta, context, data, info):
-        newcontext = GroupingContextDecorator(context)
-        newcontext.info = info
-        self.context = newcontext
-        self.meta = meta
-
-        # Call the handler contructor
-        args = toargs(newcontext, meta.schema, data)
-        self.handler = self.meta.handler(newcontext, **args)
-
-    def contained(self, name, data, info):
-        """Handle a subdirective
-        """
-        # Look up the subdirective meta data on our meta object
-        ns, name = name
-        schema = self.meta.get(name)
-        if schema is None:
-            raise ConfigurationError("Invalid directive", name)
-        schema = schema[0]  # strip off info
-        handler = getattr(self.handler, name)
-        return SimpleStackItem(self.context, handler, info, schema, data)
-
-    def finish(self):
-        # when we're done, we call the handler, which might return more actions
-        # Need to save and restore old info
-        # XXX why not just use callable()?
-        try:
-            actions = self.handler()
-        except AttributeError as v:
-            if v.args[0] == '__call__':
-                return  # noncallable
-            raise
-        except TypeError:
-            return  # non callable
-
-        if actions:
-            # we allow the handler to return nothing
-            for action in actions:
-                if not isinstance(action, dict):
-                    action = expand_action(*action)
-                self.context.action(**action)
 
 
 ##############################################################################
@@ -510,7 +359,7 @@ def toargs(context, schema, data):
 # Conflict resolution
 
 def expand_action(discriminator, callable=None, args=(), kw=None,
-                  includepath=(), info=None, order=0, **extra):
+                  order=0, **extra):
     if kw is None:
         kw = {}
     action = extra
@@ -520,8 +369,6 @@ def expand_action(discriminator, callable=None, args=(), kw=None,
             callable=callable,
             args=args,
             kw=kw,
-            includepath=includepath,
-            info=info,
             order=order))
     return action
 
@@ -531,10 +378,8 @@ def resolve_conflicts(actions):
 
     Given an actions list, identify and try to resolve conflicting actions.
     Actions conflict if they have the same non-None discriminator.
-    Conflicting actions can be resolved if the include path of one of
-    the actions is a prefix of the includepaths of the other
-    conflicting actions and is unequal to the include paths in the
-    other conflicting actions.
+    Conflicting actions can be resolved by the order they were included
+    by configured application
     """
 
     # organize actions by discriminators
@@ -574,29 +419,24 @@ def resolve_conflicts(actions):
 
     for discriminator, ainfos in unique.items():
 
-        # We use (includepath, order, i) as a sort key because we need to
-        # sort the actions by the paths so that the shortest path with a
-        # given prefix comes first.  The "first" action is the one with the
-        # shortest include path.  We break sorting ties using "order", then
-        # "i".
-        def bypath(ainfo):
-            path, order, i = ainfo[2]['includepath'], ainfo[0], ainfo[1]
-            return path, order, i
+        # We use (order, i) as a sort key because we need to
+        def byorder(ainfo):
+            order, i = ainfo[0], ainfo[1]
+            return order, i
 
-        ainfos.sort(key=bypath)
+        ainfos.sort(key=byorder)
         ainfo, rest = ainfos[0], ainfos[1:]
         output.append(ainfo)
         _, _, action = ainfo
-        basepath, baseinfo, discriminator = (action['includepath'],
-                                             action['info'],
-                                             action['discriminator'])
+        order = action['order']
+        discriminator = action['discriminator']
+        base_module_name = action['module_name']
+        base_order = action['order']
 
         for _, _, action in rest:
-            includepath = action['includepath']
-            # Test whether path is a prefix of opath
-            if (includepath[:len(basepath)] != basepath or includepath == basepath):
-                L = conflicts.setdefault(discriminator, [baseinfo])  # noqa
-                L.append(action['info'])
+            if action['order'] <= base_order:
+                L = conflicts.setdefault(discriminator, [base_module_name, base_order])
+                L.append((action['module_name'], action['order']))
 
     if conflicts:
         raise ConfigurationConflictError(conflicts)